Add determination of HydroSHEDS shapefile based on continent names
The identification of the HydroSHEDS shapefile to be used when no file has
been specified as an input argument (previously done by looking for a
HydroSHEDS bounding box that fully contains the input centroids' bounding
box) turned out to be too crude. Therefore, an additional way of
determining the HydroSHEDS shapefile based on continent name has been
implemented.