Add special case Mexico
When determinining shapefile names based on continent names, there will
be mismatches because both the naming of continents and the assignment of
the countries to the continents differs between the CLIMADA world border
shapefile and the HydroSHEDS shapefiles (e.g., world border shapefiles
contains 'Oceania' as a continent name, but would assign Mexico to North
America since it does not consider Central America a continent).
Therefore, some special cases need to be introduced manually in order to
make the dermination of the HydroSHEDS shapefile name more reliable.