Plot floods
This is just a very slightly adjusted version of climada_plot_hr.m
(the titles have been shortened since currently the flood hazard sets
do not contain a field 'datenum') that can be used to plot the floods.
This function will become redundant at some point (it doesn't make sense
to have two almost identical versions of one function), but is useful for
the moment.